Update 1.3
I've taken all of the functions that gathered information from the
database and placed them into js file that will only be used to speak to
the database. Also we can now change the main color theme and the
header/footer header text color from within the main javascript file,
this feature will be controlled from within an admin page in the future.