Update 1.4
Users can now make selections by clicking directly on the image. The
bread crumbs at the bottom of the page now have labels. The app bar that
will be used later has been moved to the top of the screen, and a few
options has been added for future use.