Update 1.5
The app is now connected to the vend POS system. The app now goes to a
thank you page when the users submits their final creation, where they
are asked if they would like to become a member. If they answer yes they
will be taken to the sign-up page which is also currently working. When
they hit submit their info is sent to both vend and the think it drink
it database for storage.